move more into new namespace, and also move the tests
