+++ src/Model/Serializer/ValueObject/Calendar/SubEventNormalizer.php
@@ -4,6 +4,7 @@ declare(strict_types=1);
 
 namespace CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar;
 
+use CultuurNet\UDB3\Model\Serializer\ValueObject\Calendar\OpeningHours\ChildcareTimeRangeNormalizer;
 use CultuurNet\UDB3\Model\Serializer\ValueObject\Contact\BookingInfoNormalizer;
 use CultuurNet\UDB3\Model\ValueObject\Calendar\SubEvent;
 use DateTimeInterface;
